DF-9997 fix UAP x64
Former-commit-id: 3a47bfbb6df7c28c4f4c8f541b5b41e998ba190b

+++ Sources/Internal/MemoryManager/MemoryManager.cpp
@@ -460,7 +460,7 @@ uint32 MemoryManager::MemorySize(void* ptr)
             return block->allocByApp;
         }
     }
-    return MallocHook::MallocSize(ptr);
+    return static_cast<uint32>(MallocHook::MallocSize(ptr));
 }
 
 void* MemoryManager::InternalAllocate(size_t size)
